Run Candlepin deploy script with QUIET=1
The deploy script of Candlepin can let the user know about the
successful end of the deployment by sending a desktop notification using
notify-send(1), if present. In Fedora 36 and greater, the installation
of the Candlepin dependencies pulls also libnotify via weak
dependencies (recommends), and that package contains notify-send(1).
The problem is that notify-send(1) sends a D-Bus message, and since
usually there is no available D-Bus bus when running the Ansible
playbook on the remote host, notify-send(1) fails with:
  Cannot autolaunch D-Bus without X11 $DISPLAY

Since we cannot notify anyway when running the deployment via Ansible,
export the QUIET=1 environment variable for the deploy script so it does
not use notify-send(1) at all. There is no command line switch for this
behaviour in any of the supported branches.